psychology_project_demo/requirements.txt
asgiref==3.4.1
dj-database-url==1.0.0
Django==3.2.15
django-heroku==0.3.1
djangorestframework==3.13.1
psycopg2
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.1.1
whitenoise==5.3.0
